+++ src/Kephas.Injection.Lite/Injection/Lite/Internal/ServiceRegistry.cs
@@ -105,11 +105,10 @@ namespace Kephas.Injection.Lite.Internal
         /// The contract declaration type is the type declaring the contract: if the <see cref="AppServiceContractAttribute.ContractType"/>
         /// is not provided, the contract declaration type is also the contract type.
         /// </summary>
-        /// <param name="context">Optional. The context in which the service types are requested.</param>
         /// <returns>
         /// An enumeration of application service information objects and their contract declaration type.
         /// </returns>
-        public IEnumerable<ContractDeclaration> GetAppServiceContracts(IContext? context = null) =>
+        public IEnumerable<ContractDeclaration> GetAppServiceContracts() =>
             this.services.Values
                 .SelectMany(s => this.ToAppServiceInfos(s).Select(si => new ContractDeclaration(si.ContractType!, si)));