+++ src/Kephas.Injection/Injection/Builder/InjectorBuilderBase.cs
@@ -210,11 +210,10 @@ namespace Kephas.Injection.Builder
             /// <summary>
             /// Gets an enumeration of application service information objects.
             /// </summary>
-            /// <param name="context">Optional. The context in which the service types are requested.</param>
             /// <returns>
             /// An enumeration of application service information objects and their associated contract type.
             /// </returns>
-            public IEnumerable<ContractDeclaration> GetAppServiceContracts(IContext? context = null)
+            public IEnumerable<ContractDeclaration> GetAppServiceContracts()
             {
                 return this.appServiceInfos.Select(i => new ContractDeclaration(i.ContractType!, i));
             }
@@ -222,11 +221,10 @@ namespace Kephas.Injection.Builder
             /// <summary>
             /// Gets an enumeration of tuples containing the service type and the contract declaration type which it implements.
             /// </summary>
-            /// <param name="context">Optional. The context in which the contracts are requested.</param>
             /// <returns>
             /// An enumeration of tuples containing the service type and the contract declaration type which it implements.
             /// </returns>
-            public IEnumerable<ServiceDeclaration> GetAppServices(IContext? context = null)
+            public IEnumerable<ServiceDeclaration> GetAppServices()
             {
                 yield break;
             }